remove forwardable module
